+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the Docker Language Server will be documented in this fil
 - Compose
   - textDocument/documentLink
     - return document links for files referenced in the short-form `volumes` attribute of a service object ([#460](https://github.com/docker/docker-language-server/issues/460))
+    - return document links for files referenced in the long-form `volumes` attribute of a service object that has a bind mount ([#462](https://github.com/docker/docker-language-server/issues/462))
 
 ## [0.18.0] - 2025-08-25