+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Docker Language Server will be documented in this file.
 
-## [Unreleased]
+## [0.19.0] - 2025-09-16
 
 ### Added
 
@@ -650,7 +650,8 @@ All notable changes to the Docker Language Server will be documented in this fil
   - textDocument/semanticTokens/full
     - provide syntax highlighting for Bake files
 
-[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.18.0...main
+[Unreleased]: https://github.com/docker/docker-language-server/compare/v0.19.0...main
+[0.19.0]: https://github.com/docker/docker-language-server/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/docker/docker-language-server/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/docker/docker-language-server/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/docker/docker-language-server/compare/v0.15.0...v0.16.0